+++ elements/compiled/dev/smarty_compiled/9492fdfdb4b2c25f886814e4bc8c7a89c39cbf42_0.file.toogle.tpl.php
@@ -1,11 +1,11 @@
 <?php
-/* Smarty version 3.1.31, created on 2018-01-29 15:46:34
+/* Smarty version 3.1.31, created on 2018-01-30 08:24:16
   from "/Applications/MAMP/htdocs/iumio-framework-se/vendor/iumio_framework/BaseApps/ManagerApp/Front/views/partials/toogle.tpl" */
 
 /* @var Smarty_Internal_Template $_smarty_tpl */
 if ($_smarty_tpl->_decodeProperties($_smarty_tpl, array (
   'version' => '3.1.31',
-  'unifunc' => 'content_5a6f33caaa0e78_96978231',
+  'unifunc' => 'content_5a701da08bd204_40639286',
   'has_nocache_code' => false,
   'file_dependency' => 
   array (
@@ -20,7 +20,7 @@ if ($_smarty_tpl->_decodeProperties($_smarty_tpl, array (
   array (
   ),
 ),false)) {
-function content_5a6f33caaa0e78_96978231 (Smarty_Internal_Template $_smarty_tpl) {
+function content_5a701da08bd204_40639286 (Smarty_Internal_Template $_smarty_tpl) {
 ?>
 <button type="button" class="navbar-toggle toggle-iumio-manager" data-toggle="collapse" data-target="#navigation-example-2" style="padding-top: 17px;">
     <span class="sr-only">Toggle navigation</span>