+++ public/setup1/setup.php
@@ -1,8 +1,9 @@
 
 <?php
 // EXIT IF DOES NOT HAVE ONE APP
-if (!empty(json_decode(file_get_contents(__DIR__."/../../elements/config_files/core/framework.config.json"))))
-    exit("iumio Installer - Cannot use iumio installer because you have already one app installed.");
+$f = json_decode(file_get_contents(__DIR__."/../../elements/config_files/core/framework.config.json"));
+if (isset($f->installation) && $f->installation != null)
+    exit("iumio Framework Installer :  Cannot use iumio installer because you have already one app installed.");
 ?>
 <!DOCTYPE html>
 <html>